+++ app/Http/Controllers/HomeController.php
@@ -8,16 +8,23 @@
 
 namespace App\Http\Controllers;
 
+use App\Gateways\PublicIPGateway;
+
 class HomeController extends Controller
 {
 
+    /**
+     * @var \App\Gateways\PublicIPGateway
+     */
+    private $gateway;
 
     /**
      * HomeController constructor.
      */
-    public function __construct()
+    public function __construct(PublicIPGateway $gateway)
     {
         $this->middleware('auth');
+        $this->setGateway($gateway);
     }
 
 
@@ -26,6 +33,23 @@ class HomeController extends Controller
      */
     public function dashboard()
     {
-        return view('v200.pages.dashboard');
+        $public_ip = $this->getGateway()->getLastRecord();
+        return view('v200.pages.dashboard')->with('public_ip', $public_ip);
+    }
+
+    /**
+     * @return \App\Gateways\PublicIPGateway
+     */
+    private function getGateway(): PublicIPGateway
+    {
+        return $this->gateway;
+    }
+
+    /**
+     * @param \App\Gateways\PublicIPGateway $gateway
+     */
+    private function setGateway(PublicIPGateway $gateway): void
+    {
+        $this->gateway = $gateway;
     }
 }