+++ resources/views/v200/pages/dashboard.blade.php
@@ -5,15 +5,28 @@
     <section class="row">
         <div class="col-sm-12">
             <section class="row">
-                <div class="col-md-12 col-lg-8">
-                    <h1>Left</h1>
-                </div>
                 <div class="col-md-12 col-lg-4">
-                    <h1>Right</h1>
+                    <div class="card text-white bg-primary">
+                        <div class="card-header">
+                            <p><i class="fa fa-wifi"></i> {{ __('common.ip_address') }}</p>
+                        </div>
+                        <div class="card-block">
+                            @php
+                                $ip_address = '-.-.-.-';
+                                $timezone = config('app.timezone');
+                                $last_update = '----';
+                            @endphp
+                            @if ($public_ip != null)
+                                @php
+                                    $ip_address = $public_ip->ip_address;
+                                    $last_update = $public_ip->updated_at;
+                                @endphp
+                            @endif
+                            <p>{{ $ip_address }}</p>
+                            <p>{{ __('common.last_update') }} {{ $last_update }}</p>
+                        </div>
+                    </div>
                 </div>
-            </section>
-            <section class="row">
-                <div class="col-12 mt-1 mb-4">Template by <a href="https://medialoot.com/preview/bootstrap-4-dashboard-premium/index.html">Medialoot</a></div>
             </section>
         </div>
     </section>