+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 # ??/??/2019
 + Add bootstrap layout
++ Add Public IP on dashboard
 
 # 0.1.0
 ## 24-03-2019